More refactoring, add the header
