+++ apps/web/src/app/admin/scenarios/new/page.tsx
@@ -1,5 +1,5 @@
+import { AdminHeader } from "@/components/admin-header";
 import { ScenarioForm } from "@/components/scenario-form/scenario-form";
-import { SiteHeader } from "@/components/site-header";
 import { getRandomScenario } from "@workspace/plantools";
 
 const defaultValues = getRandomScenario();
@@ -7,7 +7,7 @@ const defaultValues = getRandomScenario();
 export default function NewScenarioPage() {
   return (
     <div>
-      <SiteHeader />
+      <AdminHeader />
       <main className="p-4">
         <ScenarioForm defaultValues={defaultValues} />
       </main>
