+++ packages/validators/src/auth0-url.ts
@@ -1,13 +1,13 @@
 import { z } from "zod";
 
-// Normalizes input sto be an URL that work with Auth0 since they are so wildly
+// Normalizes inputs to be an URL that work with Auth0 since they are so wildly
 // inconsistent with their URL requirements.
 export const auth0url = z
   .string()
   .trim()
   .transform((value) => {
     let url = value;
-    if (!/^https?:\/\//i.test(url)) {
+    if (!/^https:\/\//i.test(url)) {
       url = `https://${url}`;
     }
     return url.replace(/\/+$/, ""); // Strip trailing slashes
